Move Global to core, but not handle_alloc_error
