+++ library/core/src/alloc/mod.rs
@@ -7,6 +7,8 @@ mod layout;
 
 #[stable(feature = "global_alloc", since = "1.28.0")]
 pub use self::global::GlobalAlloc;
+#[unstable(feature = "core_global_alloc", issue = "none")]
+pub use self::global::{Global, alloc, alloc_zeroed, dealloc, realloc};
 #[stable(feature = "alloc_layout", since = "1.28.0")]
 pub use self::layout::Layout;
 #[stable(feature = "alloc_layout", since = "1.28.0")]
