+++ tests/mir-opt/pre-codegen/drop_box_of_sized.rs
@@ -7,13 +7,13 @@
 pub unsafe fn drop_generic<T: Copy>(x: *mut Box<T>) {
     // CHECK-LABEL: fn drop_generic
     // CHECK: [[ALIGNMENT:_.+]] = const <T as std::mem::SizedTypeProperties>::ALIGN as std::mem::Alignment (Transmute)
-    // CHECK: alloc::alloc::__rust_dealloc({{.+}}, const <T as std::mem::SizedTypeProperties>::SIZE, move [[ALIGNMENT]])
+    // CHECK: core::alloc::global::__rust_dealloc({{.+}}, const <T as std::mem::SizedTypeProperties>::SIZE, move [[ALIGNMENT]])
     std::ptr::drop_in_place(x)
 }
 
 // EMIT_MIR drop_box_of_sized.drop_bytes.runtime-optimized.after.mir
 pub unsafe fn drop_bytes(x: *mut Box<[u8; 1024]>) {
     // CHECK-LABEL: fn drop_bytes
-    // CHECK: alloc::alloc::__rust_dealloc({{.+}}, const 1024_usize, {{.+}}Align1Shl0 {{.+}})
+    // CHECK: core::alloc::global::__rust_dealloc({{.+}}, const 1024_usize, {{.+}}Align1Shl0 {{.+}})
     std::ptr::drop_in_place(x)
 }
